Reorganised our components.
