+++ src/renderer/src/components/dialogs/ResetAllDialog.vue
