+++ src/renderer/src/components/propertyEditors/GraphsPropertyEditorComponent.vue
