+++ src/renderer/src/components/propertyEditors/PropertyEditorComponent.vue
