+++ src/renderer/src/components/propertyEditors/SolversPropertyEditorComponent.vue
